requirements.txt
streamlit==1.38.0
opencv-python
numpy
Pillow
streamlit-drawable-canvas